Change courses controller for using with CanCan
