--- app/helpers/courses_helper.rb
@@ -1,2 +1,0 @@
-module CoursesHelper
-end
