+++ app/views/courses/_form.html.slim
@@ -1,4 +1,4 @@
-= form_for [:admin, @course] do |f|
+= form_for [@course] do |f|
   ul.text-danger
     - @course.errors.full_messages.each do |m|
       li = m
