+++ app/views/courses/edit.html.slim
