+++ app/views/courses/new.html.slim
